+++ packages/react/src/components/TextField/index.tsx
@@ -1,8 +1,8 @@
 import React, { ChangeEvent } from 'react';
 import classNames from 'classnames';
 import rndid from '../../utils/rndid';
-import tokenList from '../../utils/token-list';
 import setRef from '../../utils/setRef';
+import { addIdRef } from '../../utils/idRefs';
 
 export interface TextFieldProps
   extends Omit<React.InputHTMLAttributes<HTMLInputElement>, 'onChange'> {
@@ -72,6 +72,7 @@ export default class TextField extends React.Component<
       requiredText,
       multiline,
       'aria-describedby': ariaDescribedby,
+      children,
       className,
       ...other
     } = this.props;
@@ -81,7 +82,7 @@ export default class TextField extends React.Component<
     const Field: any = multiline ? 'textarea' : 'input';
     const inputProps = {
       'aria-describedby': error
-        ? tokenList(this.errorId, ariaDescribedby)
+        ? addIdRef(ariaDescribedby, this.errorId)
         : ariaDescribedby
     };